functions/requirements.txt
azure-common==1.1.28
azure-core==1.27.1
azure-functions==1.15.0
azure-identity==1.13.0
azure-mgmt-compute==29.1.0
azure-mgmt-containerinstance==10.1.0
azure-mgmt-containerservice==26.0.0
azure-mgmt-core==1.4.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==23.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
idna==3.4
isodate==0.6.1
msal==1.22.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
portalocker==2.7.0
pycparser==2.21
PyJWT==2.7.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.6.3
urllib3==2.0.3
